--- src/main/java/com/bbookk/config/WebInit.java
@@ -1,29 +1,0 @@
-package com.bbookk.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.ContextLoaderListener;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.servlet.DispatcherServlet;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletRegistration;
-
-public class WebInit implements WebApplicationInitializer {
-
-    @Override
-    public void onStartup(ServletContext container)
-    {
-        AnnotationConfigWebApplicationContext context
-                = new AnnotationConfigWebApplicationContext();
-        context.setConfigLocation("com.bbookk.config");
-
-        container.addListener(new ContextLoaderListener(context));
-
-        ServletRegistration.Dynamic dispatcher = container
-                .addServlet("dispatcher", new DispatcherServlet(context));
-
-        dispatcher.setLoadOnStartup(1);
-        dispatcher.addMapping("/");
-    }
-}
