+++ src/main/java/com/bbookk/repository/BookRepository.java
@@ -3,5 +3,5 @@ package com.bbookk.repository;
 import com.bbookk.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Long> ,BookRepositoryCustom{
 }
